back-end/requirements.txt
Flask==2.3.3
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.2
mysql-connector-python==8.1.0
python-dotenv==1.0.0
